requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
# ML libraries for recommendation system
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
python-dotenv==1.0.0
